Genres update
Now the user can add the books genre and this is correctly managed when deleting or adding a book. The filter uses this genres to look for profiles. Also, the navigation bar was optimized in all windows (still pending to complete the home page). Corrected slight error when deleting/editing books. Slightly modified models (users).