Forgot password configuration
Entire frontend and backend configuration for the forgot password procedure.
NOTE: The email sending part won't work if you don't have the last updated .env content. Also, if you wanna try it out you need to register an account with an actually existing email since this is where the password resetting link is sent to.